First step in GlObject implementation working
