+++ src/dtcc_viewer/opengl/gl_mesh.py
@@ -12,6 +12,7 @@ from dtcc_viewer.opengl.environment import Environment
 from dtcc_viewer.logging import info, warning
 from dtcc_viewer.opengl.utils import Submeshes
 from dtcc_viewer.opengl.wrp_data import MeshDataWrapper
+from dtcc_viewer.opengl.gl_object import GlObject
 
 from dtcc_viewer.opengl.parameters import (
     GuiParameters,
@@ -50,7 +51,7 @@ from dtcc_viewer.shaders.shaders_color_maps import (
 )
 
 
-class GlMesh:
+class GlMesh(GlObject):
     """Represents a 3D mesh in OpenGL and provides methods for rendering.
 
     This class handles the rendering of mesh data using OpenGL. It provides methods to
@@ -106,11 +107,6 @@ class GlMesh:
     cast_shadows: bool  # If the mesh should cast shadows
     recieve_shadows: bool  # If the mesh should recieve shadows
 
-    data_texture: int  # Texture for data
-    data_wrapper: MeshDataWrapper  # Data wrapper for the mesh
-    texture_slot: int  # GL_TEXTURE0, GL_TEXTURE1, etc.
-    texture_int: int  # Texture index 0 for GL_TEXTURE0, 1 for GL_TEXTURE1, etc.
-
     def __init__(self, mesh_wrapper: MeshWrapper):
         """Initialize the MeshGL object with vertex, face, and edge information."""
 
@@ -143,11 +139,6 @@ class GlMesh:
         self.texture_slot = None
         self.texture_int = None
 
-    def preprocess(self):
-        self._create_textures()
-        self._create_geometry()
-        self._create_shaders()
-
     def get_vertex_ids(self):
         return self.vertices[8::9]
 
@@ -168,36 +159,6 @@ class GlMesh:
 
     def _create_textures(self) -> None:
         self._create_data_texture()
-
-    def _create_data_texture(self) -> None:
-        """Create texture for data."""
-
-        self.data_texture = glGenTextures(1)
-        glBindTexture(GL_TEXTURE_2D, self.data_texture)
-
-        # Configure texture filtering and wrapping options
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE)
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE)
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST)
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST)
-
-        width = self.data_wrapper.col_count
-        height = self.data_wrapper.row_count
-        key = self.data_wrapper.get_keys()[0]
-        default_data = self.data_wrapper.data_mat_dict[key]
-
-        # Transfer data to the texture
-        glTexImage2D(
-            GL_TEXTURE_2D,
-            0,
-            GL_R32F,
-            width,
-            height,
-            0,
-            GL_RED,
-            GL_FLOAT,
-            default_data,
-        )
 
     def _create_geometry(self) -> None:
         """Set up vertex and element buffers for mesh rendering."""
@@ -470,31 +431,6 @@ class GlMesh:
             self.shader_shdw, "picked_id"
         )
 
-    def _update_data_texture(self):
-        index = self.guip.data_idx
-        key = self.data_wrapper.get_keys()[index]
-        width = self.data_wrapper.col_count
-        height = self.data_wrapper.row_count
-        data = self.data_wrapper.data_mat_dict[key]
-        tic = time.perf_counter()
-
-        self._bind_data_texture()
-        glTexSubImage2D(GL_TEXTURE_2D, 0, 0, 0, width, height, GL_RED, GL_FLOAT, data)
-        self._unbind_data_texture()
-
-        toc = time.perf_counter()
-        info(f"Data texture updated. Time elapsed: {toc - tic:0.4f} seconds")
-
-    def update_color_data(self):
-        if self.guip.update_data_tex:
-            self._update_data_texture()
-            self.guip.update_data_tex = False
-
-    def update_color_caps(self):
-        if self.guip.update_caps:
-            self.guip.calc_data_min_max()
-            self.guip.update_caps = False
-
     def render_lines(
         self,
         action: Action,
@@ -688,11 +624,6 @@ class GlMesh:
         glDrawElements(GL_LINES, len(self.edges), GL_UNSIGNED_INT, None)
         self._unbind_vao()
 
-    def _bind_data_texture(self):
-        """Bind the data texture."""
-        glActiveTexture(self.texture_slot)
-        glBindTexture(GL_TEXTURE_2D, self.data_texture)
-
     def _bind_vao_triangels(self) -> None:
         """Bind the vertex array object for triangle rendering."""
         glBindVertexArray(self.VAO_triangels)
@@ -700,11 +631,6 @@ class GlMesh:
     def _bind_vao_lines(self) -> None:
         """Bind the vertex array object for wireframe rendering."""
         glBindVertexArray(self.VAO_edge)
-
-    def _unbind_data_texture(self):
-        """Unbind the currently bound data texture."""
-        glActiveTexture(self.texture_slot)
-        glBindTexture(GL_TEXTURE_2D, 0)
 
     def _unbind_vao(self) -> None:
         """Unbind the currently bound vertex array object."""
